Moves Amazon.YesIntent out to its own file and creates a new file for the handler of Amazon.NoIntent.
