+++ lib/alexa-skill/index.js
@@ -15,7 +15,9 @@ var Config                       = require('../../config'),
     request                      = require('request-promise'),
     _                            = require('lodash'),
     handleJunkPickUpDialogIntent = require('./intent-handlers/junk-pickup-dialog'),
-    handleProvideAddressIntent   = require('./intent-handlers/provide-address');
+    handleProvideAddressIntent   = require('./intent-handlers/provide-address'),
+    handleAmazonYesIntent        = require('./intent-handlers/amazon-yes'),
+    handleAmazonNoIntent        = require('./intent-handlers/amazon-yes');
 
 // ----------------------- Define the skill -----------------------
 
@@ -99,6 +101,9 @@ TheCitySkill.prototype.intentHandlers = {
     },
     'AMAZON.YesIntent': function (event, response) {
         return handleAmazonYesIntent(event, response);
+    },
+    'AMAZON.NoIntent': function (event, response) {
+        return handleAmazonNoIntent(event, response);
     }
 };
 
@@ -130,41 +135,4 @@ function handleSessionEndRequest(response) {
 
         resolve(response.tell(speechOutput, null));
     });
-}
-
-function handleAmazonYesIntent(event, response) {
-    let streetAddress = event.session.attributes.VerifiedStreetAddress;
-    let city = event.session.attributes.VerifiedCity;
-
-    if (streetAddress && city) {
-        return rubbishPickupAPI.query(streetAddress, city)
-                .then(function (result) {
-                    if (Array.isArray(result) && result.length > 0) {
-                        let speechOutput = {
-                            speech: `The junk set out will begin in your area on ${result[0].JunkSetOutBegin} and will go on till ${result[0].JunkSetOutEnd}}`,
-                            type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                        };
-
-                        return response.tell(speechOutput);
-                    } else {
-                        return new Promise(function (resolve) {
-                            let speechOutput = {
-                                speech: `Bummer! It seems to me that I don't have the junk set out data for your address. You might be out of the area covered by the city services.`,
-                                type: AlexaSkill.speechOutputType.PLAIN_TEXT
-                            };
-
-                            resolve(response.tell(speechOutput, null));
-                        });
-                    }
-                });
-    } else {
-        return new Promise(function (resolve) {
-            let speechOutput = {
-                speech: `Bummer! I'm having a technical problem that prevents me from fulfilling your request. Try again in a second.`,
-                type: AlexaSkill.speechOutputType.PLAIN_TEXT
-            };
-
-            resolve(response.tell(speechOutput, null));
-        });
-    }
 }
