+++ lib/alexa-skill/intent-handlers/amazon-no.js
@@ -1,0 +1,13 @@
+/**
+ * @module      com.louisvillemetro.the-city
+ * @author      Reydel Leon Machado
+ * @copyright   Copyright (c) 2016 Reydel Leon Machado
+ * @license     This code is licensed under the MIT license. See the LICENSE file in the project root for license terms.
+ */
+"use strict";
+
+module.exports = handleAmazonNoIntent;
+
+function handleAmazonNoIntent() {
+
+}
