+++ lib/alexa-skill/intent-handlers/junk-pickup-dialog.js
@@ -5,6 +5,8 @@
  * @license     This code is licensed under the MIT license. See the LICENSE file in the project root for license terms.
  */
 
+"use strict";
+
 module.exports = handleJunkPickUpDialogIntent;
 
 /**
@@ -15,6 +17,8 @@ module.exports = handleJunkPickUpDialogIntent;
  * @returns {Promise}
  */
 function handleJunkPickUpDialogIntent(event, response) {
+    var actions = event.session.attributes.actions;
+
     return new Promise(function (resolve) {
         let speechOutput = {
             speech: `Sure! Tell me your address, number first, including your city or state. Ignore the unit number and zip code.`,
@@ -26,6 +30,7 @@ function handleJunkPickUpDialogIntent(event, response) {
             type: AlexaSkill.speechOutputType.PLAIN_TEXT
         };
 
+        response._session.attributes.actions.push('');
         resolve(response.ask(speechOutput, repromptSpeech));
     });
 }
