+++ lib/alexa-skill/intent-handlers/provide-address.js
@@ -1,8 +1,11 @@
 /**
- * @module      com.reydelleonmachado.
+ * @module      com.louisvillemetro.the-city
+ * @author      Reydel Leon Machado
  * @copyright   Copyright (c) 2016 Reydel Leon Machado
- * @license     All rights reserved
+ * @license     This code is licensed under the MIT license. See the LICENSE file in the project root for license terms.
  */
+
+"use strict";
 
 module.exports = handleProvideAddressIntent;
